fix(#1040): address architecture and prompt-health review findings
- Move PatchStudentProfileRequest/AppendDifficultyItem to own file
- Rename ReplaceLearningGoals -> LearningGoals (neutral, action-agnostic key)
- Add validation on AppendDifficultyItem (MaxLength for Description, Competency, Subcategory)
- Extract difficulty default constants in StudentsController
- Add ModelState.IsValid guard to PatchStudentProfile
- Use unqualified ValidationException in catch block
- Remove negative-bloat clause from newStudentIntent prompt instruction
- Add non-Spanish examples to newStudentIntent to signal language-agnosticism
