+++ backend/LangTeach.Api/AI/PromptService.cs
@@ -1672,7 +1672,7 @@ public class PromptService : IPromptService
             - teachingTodoTexts: array of strings — action items or tasks for the teacher to do (e.g. "Send subjunctive exercises", "Prepare business vocabulary list"). Empty array [] if none.
             - teachingNotes: string or null — observations about how the student learns (learning style, what works, what does not). Null if not mentioned.
             - interests: array of strings — personal interests, hobbies, or topics the student enjoys. Empty array [] if not mentioned.
-            - newStudentIntent: boolean — true ONLY if the teacher explicitly states they want to create a NEW student record (e.g. "nuevo alumno", "dame de alta a", "tengo un nuevo estudiante", "registra un nuevo estudiante"). false in all other cases, including when a student name is mentioned while updating or adding data for an existing student.
+            - newStudentIntent: boolean — true ONLY if the teacher explicitly states they want to create a NEW student record (e.g. "nuevo alumno", "dame de alta a", "tengo un nuevo estudiante", "I have a new student", "register a new student").
 
             Use null for scalar fields that cannot be clearly inferred from the text.
             Keep each value concise.
